+++ apps/app/src/features/project/context/file-suggestions/index.ts
@@ -1,4 +1,9 @@
 export { FileSuggestionList } from "./FileSuggestionList";
 export type { FileSuggestion, FileSuggestionKind } from "./file-suggestions";
-export { flattenFileSuggestionTrees, matchFileSuggestions } from "./file-suggestions";
+export {
+  flattenFileSuggestionTrees,
+  folderChildren,
+  matchFileSuggestions,
+  parentPath,
+} from "./file-suggestions";
 export { useFileSuggestions } from "./use-file-suggestions";